+++ tests/integration/test_models/standard_model_tests.py
@@ -128,7 +128,8 @@ class OptimisationsTest(object):
 
         self.model = model
 
-    def evaluate_model(self, simplify=False, use_known_evals=False, to_python=False):
+    def evaluate_model(self, simplify=False, use_known_evals=False,
+                       to_python=False, to_jax=False):
         result = np.empty((0, 1))
         for eqn in [self.model.concatenated_rhs, self.model.concatenated_algebraic]:
             if simplify:
@@ -139,6 +140,9 @@ class OptimisationsTest(object):
                 eqn_eval, known_evals = eqn.evaluate(0, y, known_evals={})
             elif to_python:
                 evaluator = pybamm.EvaluatorPython(eqn)
+                eqn_eval = evaluator.evaluate(0, y)
+            elif to_jax:
+                evaluator = pybamm.EvaluatorJax(eqn)
                 eqn_eval = evaluator.evaluate(0, y)
             else:
                 eqn_eval = eqn.evaluate(0, y)